bugfix send file

+++ src/Util.cpp
@@ -37,9 +37,9 @@ namespace util {
         return send_buff(sock_fd, data.c_str(), data.length());
     }
 
-    const bool send_buff(int sock_fd, const char *buf, unsigned long size) {
-        unsigned long total = 0;
-        unsigned long bytes_remaining = size;
+    const bool send_buff(int sock_fd, const char *buf, size_t size) {
+        ssize_t total = 0;
+        size_t bytes_remaining = size;
         ssize_t n = 0;
 
         while (total < size) {
